The restoration of results did not previously ensure the correct ordering.
When reconstituting results from serialised distributions, it was possible
that the volumes were not processed in the same order as they occur in the
validation set.  As their position in that order is what connects them to
the input volumes and reference distributions, this could potentially lead
to dice and confusion metrics being applied to the wrong volumes.

As part of this effort to validate the results, the volume id is now also
required during the addition of distributions for a volume to the metrics.
This allows the append function to verify that the id matches the position
at which the results will be inserted, and hence that the results are in
the correct (validation set) order.