Corrected a possible bug in the calculation of the final volume in a batch.
The final batch in a sequential iteration will be smaller than the previous
batches if the number of patches in a batch does not evenly divide the total
number of patches in the iteration.  The code previously assume that the
difference between the full patch extent, and the number of patches available
spanned at most one volume.  While this is probably true in practice, it is
not necessarily the case, so I have generalised this to handle this case.
Still hunting down a bug which leads the predicted distribution volume to be
offset relative to the reference distribution volume.  Sadly this is not the
cause, but it is worth fixing anyway.