Corrected image extraction in the results to take account of the margin.
The Image.extract method in the results module did not previously accept
a margin parameter, only an offset.  As the predicted distribution will
cover the reconstructed shape starting at the offset + the margin, this
meant that extract would never return the same area as the predicted
distribution (unless the offsets were manually adapted to include the
marging themselves).  Sadly this still does not fix the offset bug which
affects the predicted and reference distributions (and the input input
image).