Added additional checks for output patch offsets and corrected prior checks.
This commit completes the process of making the code more robust against
errors in the calculation of patch offsets and shapes begun four commits
ago.  Having now properly rerun the experiments it would appear that one
of the three previous bugs _was_ the cause of the offset issue: it was
only an issue with updating of the output which obscured the fact that the
bug had been fixed.